usize

+++ src/misc.rs
@@ -13,10 +13,10 @@ fn col2i(s: &str) -> usize {
 }
 
 // convert board move coordinates "d2d4" to tuple form
-pub fn str2move(s: &str) -> (u8, u8) {
+pub fn str2move(s: &str) -> (usize, usize) {
     let fy: usize = s[1..2].parse::<usize>().unwrap() - 1;
     let fx: usize = col2i(&s[0..1]);
     let ty: usize = s[3..4].parse::<usize>().unwrap() - 1;
     let tx: usize = col2i(&s[2..3]);
-    ((fx * 8 + fy) as u8, (tx * 8 + ty) as u8)
+    ((fx * 8 + fy), (tx * 8 + ty))
 }
